[TASK] Adds option to enable/disable Word document format
Introduces a feature to enable or disable the generation of Word documents for proposals.
This allows administrators to control whether users can download proposals in Word format.
A new field "enabledFormatWord" has been added to the Call model and TCA configuration, along with corresponding language labels.
The download option in the frontend is now conditionally displayed based on the "enabledFormatWord" setting.